refactor(git): split exec.go into 4 focused files
